Fix quicksort
The previous quicksort implementation has failed usually at a
100k elements because of stack-overflow. For a long while I have
believed the issue to be with the algorithm, especially with the
partitioning, but I was wrong. The algorithm was syntactically
correct, but the pivot point was chosen so poorly as the last
element that the whole algorithm ran out of stack.